move all Fixed equations to the zoo
