--- docs/source/_rst/equation/equation_factory.rst
@@ -1,23 +1,0 @@
-Equation Factory
-==================
-
-.. currentmodule:: pina.equation.equation_factory
-.. autoclass:: pina._src.equation.equation_factory.FixedValue
-   :members:
-   :show-inheritance:
-
-.. autoclass::  pina._src.equation.equation_factory.FixedGradient
-   :members:
-   :show-inheritance:
-
-.. autoclass:: pina._src.equation.equation_factory.FixedFlux
-   :members:
-   :show-inheritance:
-
-.. autoclass:: pina._src.equation.equation_factory.FixedLaplacian
-   :members:
-   :show-inheritance:
-
-.. autoclass:: pina._src.equation.equation_factory.Laplace
-   :members:
-   :show-inheritance:
