+++ docs/source/_rst/equation/zoo/fixed_flux.rst
@@ -1,0 +1,7 @@
+Fixed Flux
+=====================
+.. currentmodule:: pina.equation.zoo.fixed_flux
+
+.. automodule:: pina._src.equation.zoo.fixed_flux
+    :members:
+    :show-inheritance:
