+++ docs/source/_rst/equation/zoo/fixed_gradient.rst
@@ -1,0 +1,7 @@
+Fixed Gradient
+=====================
+.. currentmodule:: pina.equation.zoo.fixed_gradient
+
+.. automodule:: pina._src.equation.zoo.fixed_gradient
+    :members:
+    :show-inheritance:
