+++ docs/source/_rst/equation/zoo/fixed_laplacian.rst
@@ -1,0 +1,7 @@
+Fixed Laplacian
+=====================
+.. currentmodule:: pina.equation.zoo.fixed_laplacian
+
+.. automodule:: pina._src.equation.zoo.fixed_laplacian
+    :members:
+    :show-inheritance:
