+++ docs/source/_rst/equation/zoo/fixed_value.rst
@@ -1,0 +1,7 @@
+Fixed Value
+=====================
+.. currentmodule:: pina.equation.zoo.fixed_value
+
+.. automodule:: pina._src.equation.zoo.fixed_value
+    :members:
+    :show-inheritance:
