+++ pina/_src/problem/zoo/acoustic_wave_problem.py
@@ -8,7 +8,8 @@ from pina._src.problem.spatial_problem import SpatialProblem
 from pina._src.condition.condition import Condition
 from pina._src.core.utils import check_consistency
 from pina._src.equation.equation import Equation
-from pina._src.equation.equation_factory import FixedValue, FixedGradient
+from pina._src.equation.zoo.fixed_value import FixedValue
+from pina._src.equation.zoo.fixed_gradient import FixedGradient
 from pina._src.equation.zoo.acoustic_wave_equation import AcousticWaveEquation
