+++ pina/_src/problem/zoo/diffusion_reaction_problem.py
@@ -3,7 +3,7 @@
 import torch
 from pina._src.condition.condition import Condition
 from pina._src.equation.equation import Equation
-from pina._src.equation.equation_factory import FixedValue
+from pina._src.equation.zoo.fixed_value import FixedValue
 from pina._src.problem.spatial_problem import SpatialProblem
 from pina._src.problem.time_dependent_problem import TimeDependentProblem
 from pina._src.core.utils import check_consistency
