+++ pina/equation/zoo.py
@@ -4,7 +4,12 @@ __all__ = [
     "AdvectionEquation",
     "AllenCahnEquation",
     "DiffusionReactionEquation",
+    "FixedFlux",
+    "FixedGradient",
+    "FixedLaplacian",
+    "FixedValue",
     "HelmholtzEquation",
+    "Laplace",
     "PoissonEquation",
     "AcousticWaveEquation",
 ]
@@ -17,3 +22,7 @@ from pina._src.equation.zoo.diffusion_reaction_equation import (
 )
 from pina._src.equation.zoo.helmholtz_equation import HelmholtzEquation
 from pina._src.equation.zoo.poisson_equation import PoissonEquation
+from pina._src.equation.zoo.fixed_value import FixedValue
+from pina._src.equation.zoo.fixed_gradient import FixedGradient
+from pina._src.equation.zoo.fixed_flux import FixedFlux
+from pina._src.equation.zoo.fixed_laplacian import FixedLaplacian, Laplace
